Refactor LocalDate to Date and added license message
