+++ src/index.js
@@ -1,5 +1,13 @@
 // @flow
-
-export {default as GraphQLLocalDate} from './localDate'
+/**
+ * Copyright (c) 2017, Dirk-Jan Rutten
+ * All rights reserved.
+ *
+ * This source code is licensed under the MIT license found in the
+ * LICENSE file in the root directory of this source tree.
+ *
+ */
+ 
+export {default as GraphQLLocalDate} from './date'
 export {default as GraphQLTime} from './time'
 export {default as GraphQLDateTime} from './dateTime'
