+++ src/time/index.js
@@ -1,4 +1,12 @@
 // @flow
+/**
+ * Copyright (c) 2017, Dirk-Jan Rutten
+ * All rights reserved.
+ *
+ * This source code is licensed under the MIT license found in the
+ * LICENSE file in the root directory of this source tree.
+ *
+ */
 
 import {GraphQLScalarType} from 'graphql'
 import type {GraphQLScalarTypeConfig} from 'graphql' // eslint-disable-line
