+++ src/utils/__tests__/formatterTests.js
@@ -1,5 +1,13 @@
 // @flow
-
+/**
+ * Copyright (c) 2017, Dirk-Jan Rutten
+ * All rights reserved.
+ *
+ * This source code is licensed under the MIT license found in the
+ * LICENSE file in the root directory of this source tree.
+ *
+ */
+ 
 import {
   serializeTime,
   serializeDate,
